+++ CHANGELOG.md
@@ -7,7 +7,9 @@
 #### Network
 - Non-constructor methods will no longer throw exceptions but return true or false upon success.
 ### New APIs
-None
+#### Events
+- Added `Nickvision::Events::Event::count()` method
+- Added `Nickvision::Events::Event::operator bool()` method
 ### Fixes
 None